User also needs ssh access to storage server
Not sure if that's a good idea, i.e., to *force* it. It means the user
can always remove the backup files, which is probably not what we want.
Otherwise it's difficult to see for the user if the sftp server is up
yet. Maybe we should create a hook into the ILO that gives a meaningful
JSON response. That would avoid all of this. Anyways, living with it for
now.